test(rocr): enable hotswap tests in rocrtst
