+++ projects/rocr-runtime/rocrtst/suites/test_common/CMakeLists.txt
@@ -451,6 +451,8 @@ aux_source_directory(${ROCRTST_ROOT}/suites/negative negativeSources)
 aux_source_directory(${ROCRTST_ROOT}/suites/stress stressSources)
 aux_source_directory(${ROCRTST_ROOT}/suites/test_common testCommonSources)
 
+list(FILTER functionalSources EXCLUDE REGEX ".*hotswap_(gfx_query|rewrite)\\.cc$")
+
 # Optional: enable NUMA async copy test (memory_async_copy_numa.cc)
 # Force disabled by default to avoid building the NUMA test source.
 set(ENABLE_COPY_NUMA OFF CACHE BOOL "Build Memory_Async_Copy_NUMA test" FORCE)
@@ -777,6 +779,102 @@ if(HAVE_HWLOC)
   target_link_libraries(${ROCRTST} ${HWLOC_TARGET})
 endif()
 
+enable_testing()
+
+get_filename_component(HSA_RUNTIME_ROOT "${ROCRTST_ROOT}/../runtime/hsa-runtime" ABSOLUTE)
+
+foreach(_prefix IN LISTS CMAKE_PREFIX_PATH)
+  if(EXISTS "${_prefix}/tools/comgr/lib/cmake/amd_comgr/amd_comgr-config.cmake")
+    set(amd_comgr_DIR "${_prefix}/tools/comgr/lib/cmake/amd_comgr"
+      CACHE PATH "Path to amd_comgr config" FORCE)
+    break()
+  endif()
+endforeach()
+find_package(amd_comgr CONFIG REQUIRED)
+if(TARGET amd_comgr)
+  set(HOTSWAP_COMGR_TARGET amd_comgr)
+elseif(TARGET amd_comgr::amd_comgr)
+  set(HOTSWAP_COMGR_TARGET amd_comgr::amd_comgr)
+else()
+  message(FATAL_ERROR "amd_comgr package found without an amd_comgr target")
+endif()
+
+set(HOTSWAP_GFX1250_ASM "${ROCRTST_ROOT}/suites/functional/fixtures/gfx1250_min.s")
+set(HOTSWAP_GFX1250_HSACO "${PROJECT_BINARY_DIR}/gfx1250_min.hsaco")
+set(HOTSWAP_GFX1250_HSACO_HDR "${PROJECT_BINARY_DIR}/gfx1250_min_hsaco.h")
+set(HOTSWAP_GENERATE_HSACO_HDR
+    "${ROCRTST_ROOT}/suites/functional/fixtures/generate_hsaco_header.cmake")
+
+add_custom_command(
+  OUTPUT "${HOTSWAP_GFX1250_HSACO}"
+  COMMAND ${CLANG}
+          -x assembler
+          -target amdgcn-amd-amdhsa
+          -mcpu=gfx1250
+          -o "${HOTSWAP_GFX1250_HSACO}"
+          "${HOTSWAP_GFX1250_ASM}"
+  DEPENDS "${HOTSWAP_GFX1250_ASM}" ${CLANG}
+  COMMENT "BUILDING ${HOTSWAP_GFX1250_HSACO}"
+  VERBATIM)
+
+add_custom_command(
+  OUTPUT "${HOTSWAP_GFX1250_HSACO_HDR}"
+  COMMAND ${CMAKE_COMMAND}
+          -DINPUT="${HOTSWAP_GFX1250_HSACO}"
+          -DOUTPUT="${HOTSWAP_GFX1250_HSACO_HDR}"
+          -P "${HOTSWAP_GENERATE_HSACO_HDR}"
+  DEPENDS "${HOTSWAP_GFX1250_HSACO}" "${HOTSWAP_GENERATE_HSACO_HDR}"
+  COMMENT "GENERATING ${HOTSWAP_GFX1250_HSACO_HDR}"
+  VERBATIM)
+
+add_custom_target(hotswap_gfx1250_fixture DEPENDS "${HOTSWAP_GFX1250_HSACO_HDR}")
+
+add_executable(hotswap_gfx_query
+  ${ROCRTST_ROOT}/suites/functional/hotswap_gfx_query.cc
+  ${HSA_RUNTIME_ROOT}/core/runtime/hotswap_gfx_query.cpp)
+target_include_directories(hotswap_gfx_query PRIVATE
+  ${HSA_RUNTIME_ROOT}
+  ${ROCRTST_ROOT}/gtest/include)
+target_link_libraries(hotswap_gfx_query ${GOOGLE_TEST_FRWK_NAME} pthread)
+add_test(NAME hotswap_gfx_query COMMAND hotswap_gfx_query)
+set_tests_properties(hotswap_gfx_query PROPERTIES
+  LABELS "quick;standard;pr;pre-commit;hotswap"
+  TIMEOUT 60)
+
+add_executable(hotswap_rewrite
+  ${ROCRTST_ROOT}/suites/functional/hotswap_rewrite.cc
+  ${HSA_RUNTIME_ROOT}/core/runtime/hotswap.cpp
+  ${HSA_RUNTIME_ROOT}/core/runtime/hotswap_gfx_query.cpp
+  "${HOTSWAP_GFX1250_HSACO_HDR}")
+add_dependencies(hotswap_rewrite hotswap_gfx1250_fixture)
+target_include_directories(hotswap_rewrite PRIVATE
+  ${HSA_RUNTIME_ROOT}
+  ${PROJECT_BINARY_DIR}
+  ${ROCRTST_ROOT}/gtest/include)
+target_compile_definitions(hotswap_rewrite PRIVATE ROCR_HOTSWAP_TESTING)
+target_link_libraries(hotswap_rewrite ${GOOGLE_TEST_FRWK_NAME} ${CMAKE_DL_LIBS} pthread)
+set_target_properties(hotswap_rewrite PROPERTIES
+  LINK_FLAGS "-Wl,--enable-new-dtags"
+  BUILD_RPATH "$<TARGET_FILE_DIR:${HOTSWAP_COMGR_TARGET}>;$ORIGIN"
+  INSTALL_RPATH "$ORIGIN/../lib;$ORIGIN/../lib64")
+if(UNIX)
+  add_custom_command(TARGET hotswap_rewrite POST_BUILD
+    COMMAND ${CMAKE_COMMAND} -E remove -f
+      "$<TARGET_FILE_DIR:hotswap_rewrite>/libamd_comgr.so.3"
+      "$<TARGET_FILE_DIR:hotswap_rewrite>/libamd_comgr.so"
+    COMMAND ${CMAKE_COMMAND} -E create_symlink
+      "$<TARGET_FILE:${HOTSWAP_COMGR_TARGET}>"
+      "$<TARGET_FILE_DIR:hotswap_rewrite>/libamd_comgr.so.3"
+    COMMAND ${CMAKE_COMMAND} -E create_symlink
+      "libamd_comgr.so.3"
+      "$<TARGET_FILE_DIR:hotswap_rewrite>/libamd_comgr.so"
+    VERBATIM)
+endif()
+add_test(NAME hotswap_rewrite COMMAND hotswap_rewrite)
+set_tests_properties(hotswap_rewrite PROPERTIES
+  LABELS "quick;standard;pr;pre-commit;hotswap"
+  TIMEOUT 120)
+
 #Build kernels
 add_custom_target(rocrtst_kernels ALL DEPENDS ${HSACO_TARG_LIST})
 
@@ -794,6 +892,9 @@ set_target_properties(${ROCRTST} PROPERTIES
 install(TARGETS ${ROCRTST}
         ARCHIVE DESTINATION lib
         LIBRARY DESTINATION lib
+        RUNTIME DESTINATION bin)
+
+install(TARGETS hotswap_gfx_query hotswap_rewrite
         RUNTIME DESTINATION bin)
 
 # Install platform configuration file
@@ -840,6 +941,16 @@ if(EXISTS "${TEST_CATEGORIES_CMAKE}" AND EXISTS "${TEST_CATEGORIES_YAML}")
       "${TEST_CATEGORIES_YAML}"
       "${PROJECT_BINARY_DIR}"
       "${INSTALL_TEST_FILE}"
+    )
+
+    file(
+        APPEND
+        "${INSTALL_TEST_FILE}"
+        "\n"
+        "add_test(hotswap_gfx_query \"../hotswap_gfx_query\")\n"
+        "set_tests_properties(hotswap_gfx_query PROPERTIES LABELS \"quick;standard;pr;pre-commit;hotswap\" TIMEOUT 60)\n\n"
+        "add_test(hotswap_rewrite \"../hotswap_rewrite\")\n"
+        "set_tests_properties(hotswap_rewrite PROPERTIES LABELS \"quick;standard;pr;pre-commit;hotswap\" TIMEOUT 120)\n\n"
     )
 
     install(
