+++ projects/rocr-runtime/runtime/docs/data/env_variables.rst
@@ -93,6 +93,18 @@
       - | 0: Disable debug mode.
         | 1: Enable debug mode with additional validation and logging.
 
+    * - | ``HSA_HOTSWAP_DISABLE``
+        | Disables HotSwap code object rewriting. HotSwap is enabled by default on supported GPU and ASIC revisions.
+      - ``0``
+      - | 0, ``false``, ``off``, ``no``, ``n``, or ``f``: Allow HotSwap on supported GPU and ASIC revisions.
+        | Any other non-empty value: Disable HotSwap and use the original code object load path.
+
+    * - | ``HSA_HOTSWAP_VERBOSE``
+        | Enables HotSwap diagnostic logging to stderr.
+      - ``0``
+      - | 0, ``false``, ``off``, ``no``, ``n``, or ``f``: Disable HotSwap diagnostic logging.
+        | Any other non-empty value: Enable HotSwap diagnostic logging.
+
     * - | ``HSA_ENABLE_DXG_DETECTION``
         | Controls detection of the DXG driver (/dev/dxg) on WSL2.
       - ``1``
